fix: TM-452 rename entry format function
