+++ src/admin/components/ResourceTabs/InformationTab/components/InformationTabRow.tsx
@@ -2,7 +2,7 @@ import { Divider, Typography } from "@mui/material";
 import { LabeledClasses } from "react-admin";
 import { Else, If, Then, When } from "react-if";
 
-import { convertDateFormat } from "@/admin/apiProvider/utils/dateFormat";
+import { formatEntryValue } from "@/admin/apiProvider/utils/entryFormat";
 import List from "@/components/extensive/List/List";
 import { FormSummaryRowProps, useGetFormEntries } from "@/components/extensive/WizardForm/FormSummaryRow";
 
@@ -27,9 +27,9 @@ const InformationTabRow = ({ index, ...props }: FormSummaryRowProps) => {
             </Typography>
             <If condition={typeof entry.value === "string" || typeof entry.value === "number"}>
               <Then>
-                <Typography variant="body2" dangerouslySetInnerHTML={{ __html: convertDateFormat(entry.value) }} />
+                <Typography variant="body2" dangerouslySetInnerHTML={{ __html: formatEntryValue(entry.value) }} />
               </Then>
-              <Else>{convertDateFormat(entry.value)}</Else>
+              <Else>{formatEntryValue(entry.value)}</Else>
             </If>
           </div>
         )}
