+++ src/admin/modules/application/components/ApplicationTabs.tsx
@@ -10,7 +10,7 @@ import {
 } from "react-admin";
 import { Else, If, Then, When } from "react-if";
 
-import { convertDateFormat } from "@/admin/apiProvider/utils/dateFormat";
+import { formatEntryValue } from "@/admin/apiProvider/utils/entryFormat";
 import List from "@/components/extensive/List/List";
 import { FormSummaryRowProps, useGetFormEntries } from "@/components/extensive/WizardForm/FormSummaryRow";
 import { ApplicationRead, FormSubmissionRead } from "@/generated/apiSchemas";
@@ -33,9 +33,9 @@ const ApplicationTabRow = ({ index, ...props }: FormSummaryRowProps) => {
             </Typography>
             <If condition={typeof entry.value === "string" || typeof entry.value === "number"}>
               <Then>
-                <Typography variant="body2" dangerouslySetInnerHTML={{ __html: convertDateFormat(entry.value) }} />
+                <Typography variant="body2" dangerouslySetInnerHTML={{ __html: formatEntryValue(entry.value) }} />
               </Then>
-              <Else>{convertDateFormat(entry.value)}</Else>
+              <Else>{formatEntryValue(entry.value)}</Else>
             </If>
           </div>
         )}
