+++ src/components/extensive/WizardForm/FormSummaryRow.tsx
@@ -4,7 +4,7 @@ import dynamic from "next/dynamic";
 import { useMemo } from "react";
 import { Else, If, Then } from "react-if";
 
-import { convertDateFormat } from "@/admin/apiProvider/utils/dateFormat";
+import { formatEntryValue } from "@/admin/apiProvider/utils/entryFormat";
 import Accordion from "@/components/elements/Accordion/Accordion";
 import { getDisturbanceTableColumns } from "@/components/elements/Inputs/DataTable/RHFDisturbanceTable";
 import { getFundingTypeTableColumns } from "@/components/elements/Inputs/DataTable/RHFFundingTypeDataTable";
@@ -176,10 +176,10 @@ const FormSummaryRow = ({ step, index, ...props }: FormSummaryRowProps) => {
             <If condition={typeof entry.value === "string" || typeof entry.value === "number"}>
               <Then>
                 <Text variant="text-body-300" className="flex-1" containHtml>
-                  {convertDateFormat(entry.value)}
+                  {formatEntryValue(entry.value)}
                 </Text>
               </Then>
-              <Else>{convertDateFormat(entry.value)}</Else>
+              <Else>{formatEntryValue(entry.value)}</Else>
             </If>
           </div>
         )}
